refactor(library/Sequence): moves "Byte/Encoded/Compatibility/exports.ts" to "Byte/Encoded/Compatibility/ensure.ts"
