+++ src/library/Sequence/Byte/Encoded/Compatibility/ensure.ts
