+++ src/library/Sequence/Byte/Encoded/Compatibility/index.ts
@@ -1,1 +1,1 @@
-export * as Sequence_Byte_Encoded_Compatibility from './exports.ts';
+export * as Sequence_Byte_Encoded_Compatibility from './ensure.ts';
